requirements.txt
python-dotenv==1.0.0
python-Levenshtein==0.21.1
mpstemmer @ git+https://github.com/ariaghora/mpstemmer.git
PyYAML==6.0.1
progress==1.6
numpy==1.22.4
scikit-learn==1.3.0
matplotlib==3.7.2
pandas==1.5.3
jqmcvi @ git+https://github.com/jqmviegas/jqm_cvi.git
kneed==0.8.5
